Fix reroot function.
There was a bug in `get_path_to_root` function that returned the baltic super-root, which was being accounted for in `reroot` before. No longer needed, so removed.
There was a rare error when rerooting some trees where exceptions would be raised if `fixSingletons` was set to True. Now seems to be fixed.