WIP Parallel rooting by regression with uncertain tip dates
`branchLike` now also contains a `absoluteTimeRange` parameter that can be automatically set for tips with uncertain dates with updates to `io`.
When calling `root_by_regression`, parallel workers each get a root and if there are tips with uncertain dates, those get sampled using Monte Carlo. Might want to make this optional in the future.